Create main.cpp and exclude it from the coverage testing
