+++ codecov.yml
@@ -1,3 +1,4 @@
 coverage:
   ignore:
     - "test/*"
+    - "src/main.cpp"
