+++ test/CMakeLists.txt
@@ -16,10 +16,10 @@ if (COVERAGE)
     add_custom_target(clean-coverage COMMAND find ${CMAKE_BINARY_DIR} -name '*.gcda' -exec rm {} \\\; )
     add_custom_target(tests-cov
         DEPENDS clean-coverage tests-run
-        COMMAND gcovr -r ${CMAKE_SOURCE_DIR}/src ${CMAKE_BINARY_DIR})
+        COMMAND gcovr -r ${CMAKE_SOURCE_DIR}/src ${CMAKE_BINARY_DIR} -e main.cpp)
     add_custom_target(tests-cov-html
         DEPENDS clean-coverage tests-run
-        COMMAND gcovr -r ${CMAKE_SOURCE_DIR}/src ${CMAKE_BINARY_DIR} --html --html-details -o tests.html)
+        COMMAND gcovr -r ${CMAKE_SOURCE_DIR}/src ${CMAKE_BINARY_DIR} -e main.cpp --html --html-details -o tests.html)
 endif()
 
 add_custom_target(tests-valgrind
